Signals and refactory
Changes:
- add new signals and reorganize signals
- refactory of _bulk_get_or_create to BulkOperation
- optimize deletion
