+++ edgy/core/db/querysets/mixins/__init__.py
@@ -1,11 +1,9 @@
 from edgy.core.db.context_vars import set_schema, with_schema
 
-from .bulk import BulkMixin
 from .queryset_props import QuerySetPropsMixin
 from .tenancy import TenancyMixin, activate_schema, deactivate_schema
 
 __all__ = [
-    "BulkMixin",
     "QuerySetPropsMixin",
     "TenancyMixin",
     "activate_schema",
