+++ edgy/core/db/relationships/relation.py
@@ -1,12 +1,14 @@
 from __future__ import annotations
 
 from collections.abc import Sequence
+from functools import cached_property
 from typing import TYPE_CHECKING, Any, Literal, cast
 
 from pydantic import BaseModel
 
-from edgy.core.db.context_vars import CURRENT_INSTANCE
 from edgy.core.db.fields.base import RelationshipField
+from edgy.core.db.querysets.bulk import BulkOperation
+from edgy.core.db.querysets.clauses import and_, or_
 from edgy.exceptions import ObjectNotFound, RelationshipIncompatible, RelationshipNotFound
 from edgy.protocols.many_relationship import ManyRelationProtocol
 
@@ -80,6 +82,13 @@ class ManyRelation(ManyRelationProtocol):
         # Stage the initial references.
         self.stage(*refs)
 
+    @cached_property
+    def _shared_relation_signals_params(self) -> dict:
+        fk = self.through.meta.fields[self.from_foreign_key]
+        source = fk.target
+        fk_source = source.meta.fields[fk.reverse_name]
+        return {"source": source, "field": fk_source.name, "relation": "many_to_many"}
+
     def get_queryset(self) -> QuerySet:
         """
         Returns a `QuerySet` for fetching related instances through the `through` model.
@@ -115,7 +124,7 @@ class ManyRelation(ManyRelationProtocol):
         if self.embed_through != "":
             queryset.embed_parent_filters = queryset.embed_parent
         if self.reverse:
-            if not self.through.meta.fields[self.from_foreign_key].is_cross_db():
+            if not fk.is_cross_db():
                 # not initialized yet
                 queryset._select_related.add(self.from_foreign_key)
         else:
@@ -134,20 +143,22 @@ class ManyRelation(ManyRelationProtocol):
         self.refs.clear()
         if not refs:
             return
-        await self.get_queryset()._bulk_get_update_or_create(
-            objs=refs,
+        await BulkOperation(
+            owner=self.get_queryset(),
             unique_columns={
                 col
                 for field in [self.from_foreign_key, self.to_foreign_key]
                 for col in self.through.meta.field_to_column_names[field]
             },
             ignore_create_conflicts=True,
+            signal_postfix="relation",
+            signal_params={"operation": "save_related", **self._shared_relation_signals_params},
             create=True,
             retrieve=False,
             update=False,
             used_instance=self.instance,
             resolve_embed=False,
-        )
+        )(refs)
 
     def __getattr__(self, item: Any) -> Any:
         """
@@ -189,7 +200,7 @@ class ManyRelation(ManyRelationProtocol):
         # get_queryset already returns a fresh queryset, so no need to make a copy.
         return self.get_queryset()
 
-    def expand_relationship(self, value: Any) -> Any:
+    def expand_relationship(self, value: Any) -> BaseModelType:
         """
         Expands a given value into an instance of the `through` model or its
         proxy model, preparing it for inclusion in the relationship. This
@@ -220,7 +231,7 @@ class ManyRelation(ManyRelationProtocol):
 
         # Create a new proxy model instance of the 'through' model.
         # This instance links the current 'from' model instance with the 'to' model instance.
-        instance = cast("type[BaseModelType]", through.proxy_model)(
+        instance = through.proxy_model(
             **{self.from_foreign_key: self.instance, self.to_foreign_key: value}
         )
         # Set identifying database fields for the 'through' model instance.
@@ -292,20 +303,22 @@ class ManyRelation(ManyRelationProtocol):
             "list[BaseModelType | None]",
             [
                 tup[0]
-                for tup in await self.get_queryset()._bulk_get_update_or_create(
-                    objs=prepared,
+                for tup in await BulkOperation(
+                    owner=self.get_queryset(),
                     unique_columns={
                         col
                         for field in [self.from_foreign_key, self.to_foreign_key]
                         for col in through.meta.field_to_column_names[field]
                     },
                     ignore_create_conflicts=True,
+                    signal_postfix="relation",
+                    signal_params={"operation": "add", **self._shared_relation_signals_params},
                     create=True,
                     retrieve=False,
                     update=False,
                     used_instance=self.instance,
                     resolve_embed=True,
-                )
+                )(prepared)
             ],
         )
 
@@ -341,8 +354,53 @@ class ManyRelation(ManyRelationProtocol):
             RelationshipNotFound: If no child is found or specified for removal.
             RelationshipIncompatible: If a child type is not compatible.
         """
-        for child in children:
-            await self.remove(child)
+
+        def _helper_prepare(child: Any) -> Any:
+            # Validate that the child is not a dict.
+            if isinstance(child, dict):
+                # this fails later and is wanted
+                child = None
+            return self.expand_relationship(child)
+
+        prepared = [_helper_prepare(child) for child in children]
+        through = self.through.get_real_class()
+        fk = self.through.meta.fields[self.from_foreign_key]
+        fk_source = fk.target.meta.fields[fk.reverse_name]
+        model_based_deletion = (
+            fk.use_model_based_deletion or through.__require_model_based_deletion__
+        )
+        await through.meta.signals.pre_relation.send_async(
+            through,
+            instance=self.instance,
+            values=prepared,
+            operation="delete",
+            model_based_deletion=model_based_deletion,
+            **self._shared_relation_signals_params,
+        )
+        queryset = self.get_queryset()
+        clauses = [and_(*child.identifying_clauses()) for child in prepared]
+        expression = queryset.filter(or_(*clauses))
+        async with queryset.transaction():
+            row_count = await expression.delete(model_based_deletion)
+            if row_count is not None and row_count != len(clauses):
+                related_name = fk_source.name
+                raise RelationshipNotFound(
+                    detail=(
+                        f"There is no relationship through '{related_name}' to {self.instance} from "
+                        f"{('one of ' if len(clauses) > 1 else '')}"
+                        f"{', '.join(str(getattr(child, self.to_foreign_key)) for child in prepared)}."
+                    )
+                )
+        await through.meta.signals.post_relation.send_async(
+            through,
+            instance=self.instance,
+            values=prepared,
+            field=self.from_foreign_key,
+            operation="delete_many",
+            relation="many_to_many",
+            row_count=row_count,
+            model_based_deletion=model_based_deletion,
+        )
 
     async def remove(self, child: BaseModelType | None = None) -> None:
         """
@@ -358,13 +416,13 @@ class ManyRelation(ManyRelationProtocol):
             RelationshipNotFound: If no child is found or specified for removal.
             RelationshipIncompatible: If the child type is not compatible.
         """
-        # Determine the foreign key based on whether it's a reverse relationship.
-        if self.reverse:
-            fk = self.through.meta.fields[self.from_foreign_key]
-        else:
-            fk = self.through.meta.fields[self.to_foreign_key]
 
         if child is None:
+            # Determine the foreign key based on whether it's a reverse relationship.
+            if self.reverse:
+                fk = self.through.meta.fields[self.from_foreign_key]
+            else:
+                fk = self.through.meta.fields[self.to_foreign_key]
             # If no child is specified and the foreign key is unique, attempt to get a single child.
             if fk.unique:
                 try:
@@ -375,31 +433,7 @@ class ManyRelation(ManyRelationProtocol):
             else:
                 # If no child is specified and the foreign key is not unique, raise an error.
                 raise RelationshipNotFound(detail="No child specified.")
-
-        # Cast the child to BaseModelType as it is now confirmed to be a model instance.
-        child = cast("BaseModelType", self.expand_relationship(child))
-        # Count the number of relationships based on the identifying clauses of the child.
-        real_class = self.get_real_class()
-        await child.meta.signals.pre_delete.send_async(
-            real_class, instance=child, model_instance=child
-        )
-        token = CURRENT_INSTANCE.set(child)
-        try:
-            row_count = await child.raw_delete(
-                skip_post_delete_hooks=False,
-                remove_referenced_call=False,
-            )
-            if row_count == 0:
-                # If no relationship is found, raise an error.
-                raise RelationshipNotFound(
-                    detail=f"There is no relationship between '{self.from_foreign_key}' and "
-                    f"'{self.to_foreign_key}: {getattr(child, self.to_foreign_key).pk}'."
-                )
-        finally:
-            CURRENT_INSTANCE.reset(token)
-            await self.meta.signals.post_delete.send_async(
-                real_class, instance=child, model_instance=child, row_count=row_count
-            )
+        await self.remove_many(cast("BaseModelType", child))
 
     def __repr__(self) -> str:
         """
@@ -482,6 +516,12 @@ class SingleRelation(ManyRelationProtocol):
             refs = [refs]
         # Stage the initial references.
         self.stage(*refs)
+
+    @cached_property
+    def _shared_relation_signals_params(self) -> dict:
+        fk = self.to.meta.fields[self.to_foreign_key]
+        # here reverse_name=related_name
+        return {"source": self.to, "field": fk.reverse_name, "relation": "one_to_many"}
 
     def get_queryset(self) -> QuerySet:
         """
@@ -571,7 +611,7 @@ class SingleRelation(ManyRelationProtocol):
             setattr(value, self.to_foreign_key, self.instance)
             return value
 
-        related_columns = self.to.meta.fields[self.to_foreign_key].related_columns.keys()
+        related_columns = tuple(self.to.meta.fields[self.to_foreign_key].related_columns.keys())
         # If there's only one related column and the value is not a dict or BaseModel,
         # wrap it in a dictionary with the related column name as key.
         if len(related_columns) == 1 and not isinstance(value, dict | BaseModel):
@@ -580,7 +620,7 @@ class SingleRelation(ManyRelationProtocol):
         target_instance = target.proxy_model(**value)
         setattr(target_instance, self.to_foreign_key, self.instance)
         # Set identifying database fields for the 'to' model instance.
-        target_instance.identifying_db_fields = related_columns  # type: ignore
+        target_instance.identifying_db_fields = related_columns
         # If the 'to' model is a tenant model, set the active schema for the instance.
         if getattr(target.meta, "is_tenant", False):
             target_instance.__using_schema__ = self.instance.get_active_instance_schema()  # type: ignore
@@ -640,17 +680,19 @@ class SingleRelation(ManyRelationProtocol):
         if not refs:
             return
         to = self.to
-        await self.get_queryset()._bulk_get_update_or_create(
-            objs=refs,
+        await BulkOperation(
+            owner=self.get_queryset(),
             unique_columns=to.pkcolumns,
             update_fields={self.to_foreign_key},
+            signal_postfix="relation",
+            signal_params={"operation": "save_related", **self._shared_relation_signals_params},
             ignore_create_conflicts=True,
             retrieve=False,
             create=True,
             update=True,
             used_instance=self.instance,
             resolve_embed=False,
-        )
+        )(refs)
 
     async def create(self, *args: Any, **kwargs: Any) -> BaseModelType | None:
         """
@@ -710,17 +752,19 @@ class SingleRelation(ManyRelationProtocol):
             "list[BaseModelType | None]",
             [
                 tup[0]
-                for tup in await self.get_queryset()._bulk_get_update_or_create(
-                    objs=prepared,
+                for tup in await BulkOperation(
+                    owner=self.get_queryset(),
                     unique_columns=to.pkcolumns,
                     update_fields={self.to_foreign_key},
+                    signal_postfix="relation",
+                    signal_params={"operation": "add", **self._shared_relation_signals_params},
                     ignore_create_conflicts=True,
                     retrieve=False,
                     create=True,
                     update=True,
                     used_instance=self.instance,
                     resolve_embed=True,
-                )
+                )(prepared)
             ],
         )
 
@@ -739,8 +783,42 @@ class SingleRelation(ManyRelationProtocol):
             RelationshipNotFound: If no child is found or specified for removal.
             RelationshipIncompatible: If a child type is not compatible.
         """
-        for child in children:
-            await self.remove(child)
+
+        def _helper_prepare(child: Any) -> Any:
+            # Validate that the child is not a dict.
+            if isinstance(child, dict):
+                # this fails later and is wanted
+                child = None
+            child = self.expand_relationship(child)
+            setattr(child, self.to_foreign_key, None)
+            return child
+
+        to = self.to
+        queryset = self.get_queryset()
+        operation = BulkOperation(
+            owner=queryset,
+            unique_columns=to.pkcolumns,
+            update_fields={self.to_foreign_key},
+            signal_postfix="relation",
+            signal_params={"operation": "delete", **self._shared_relation_signals_params},
+            retrieve=False,
+            create=False,
+            update=True,
+            used_instance=self.instance,
+            resolve_embed=False,
+        )
+        prepared = [_helper_prepare(child) for child in children]
+        async with queryset.transaction():
+            await operation(prepared)
+            if operation.row_count_update != len(operation.instances_and_created):
+                related_name = self.to.meta.fields[self.to_foreign_key].related_name
+                raise RelationshipNotFound(
+                    detail=(
+                        f"There is no relationship through '{related_name}' to {self.instance} from "
+                        f"{('one of ' if len(prepared) > 1 else '')}"
+                        f"{', '.join(str(child) for child in prepared)}."
+                    )
+                )
 
     async def remove(self, child: BaseModelType | None = None) -> None:
         """
@@ -769,19 +847,7 @@ class SingleRelation(ManyRelationProtocol):
             else:
                 # If no child is specified and the foreign key is not unique, raise an error.
                 raise RelationshipNotFound(detail="no child specified")
-        # Validate that the child is an instance of the 'to' model.
-        if not isinstance(child, self.to):
-            raise RelationshipIncompatible(f"The child is not from the type '{self.to.__name__}'.")
-
-        # Save the child, setting its foreign key to None to remove the relationship.
-        # MAYBE: optimize the case if the foreign_key is not matching.
-        # We need to check if the model is loaded completely
-
-        token = CURRENT_INSTANCE.set(child)
-        try:
-            await child.real_save(force_insert=False, values={self.to_foreign_key: None})
-        finally:
-            CURRENT_INSTANCE.reset(token)
+        await self.remove_many(cast("BaseModelType", child))
 
     def __repr__(self) -> str:
         """
