+++ tests/foreign_keys/test_many_to_many_field.py
@@ -173,7 +173,7 @@ async def test_raises_RelationshipNotFound(create_test_database):
 
     assert (
         raised.value.args[0]
-        == f"There is no relationship between 'album' and 'track: {track3.pk}'."
+        == f"There is no relationship through 'tracks' to {album} from {track3}."
     )
